Split up the modpanel controller
It was getting a bit large
