+++ .github/workflows/ci-pipeline.yml
@@ -49,12 +49,18 @@ jobs:
         && cd ../
 
     # Run all tests
-    - name: Run normal backend tests
+    - name: Run unit tests
       run: >
         dotnet test
         --no-build
         --verbosity normal
-        --filter FullyQualifiedName\!~Draughts.IntegrationTest.EndToEnd.Database
+        --filter FullyQualifiedName~Draughts.Test
+    - name: Run in-memory backend tests
+      run: >
+        dotnet test
+        --no-build
+        --verbosity normal
+        --filter FullyQualifiedName~Draughts.IntegrationTest.EndToEnd.InMemory
     - name: Run database backend tests
       run: >
         dotnet test
